Modify the design
